+++ lib/src/features/listener/flutter_ddi_stream_listener.dart
@@ -1,0 +1,33 @@
+import 'package:dart_ddi/dart_ddi.dart';
+import 'package:flutter/widgets.dart';
+
+/// Abstract class representing a stream listener.
+/// This mixin should be used when you want to listen for a stream.
+/// [StreamListenType] The type of the stream to listen.
+mixin StreamListener<StateType extends StatefulWidget,
+    StreamListenType extends Object> on State<StateType> {
+  StreamListenType? _state;
+  StreamListenType? get state => _state;
+
+  void _onEvent(StreamListenType listen) {
+    setState(() {
+      _state = listen;
+    });
+  }
+
+  @protected
+  @mustCallSuper
+  @override
+  void initState() {
+    ddiStream.subscribe<StreamListenType>(callback: _onEvent);
+    super.initState();
+  }
+
+  @protected
+  @mustCallSuper
+  @override
+  void dispose() {
+    ddiStream.close<StreamListenType>();
+    super.dispose();
+  }
+}