+++ lib/src/features/state/application_state.dart
@@ -3,7 +3,8 @@ import 'package:flutter/widgets.dart';
 
 /// Abstract class representing the application state.
 /// This class should be extended when you want to register and manage dependencies with DDI.
-abstract class ApplicationState<StateType extends StatefulWidget, BeanType extends Object> extends State<StateType> {
+abstract class ApplicationState<StateType extends StatefulWidget,
+    BeanType extends Object> extends State<StateType> {
   /// The `clazzRegister` function should return the instance of the bean to be registered.
   ApplicationState(BeanType Function() clazzRegister) {
     _isModule = clazzRegister is DDIModule Function();