+++ lib/src/features/state/dependent_state.dart
@@ -3,8 +3,8 @@ import 'package:flutter/widgets.dart';
 
 /// Abstract class representing a dependent state.
 /// This class should be extended when you want to register and manage dependent dependencies with DDI.
-/// Be Aware about registering the same instances more than once
-abstract class DependentState<StateType extends StatefulWidget, BeanType extends Object> extends State<StateType> {
+abstract class DependentState<StateType extends StatefulWidget,
+    BeanType extends Object> extends State<StateType> {
   /// The `clazzRegister` function should return the instance of the bean to be registered.
   DependentState(BeanType Function() clazzRegister) {
     _isModule = clazzRegister is DDIModule Function();